+++ cli/crates/server/src/bridge/server.rs
@@ -6,9 +6,9 @@ use crate::bridge::log::log_event_endpoint;
 use crate::bridge::search::search_endpoint;
 use crate::bridge::types::{Constraint, ConstraintKind, OperationKind};
 use crate::bridge::udf::invoke_udf_endpoint;
+use crate::config::DetectedUdf;
 use crate::errors::ServerError;
 use crate::event::{wait_for_event, Event};
-use crate::servers::DetectedUdf;
 use crate::types::ServerMessage;
 use axum::extract::State;
 use axum::Json;
