+++ cli/crates/server/src/lib.rs
@@ -24,6 +24,7 @@ extern crate log;
 mod atomics;
 mod bridge;
 mod codegen_server;
+mod config;
 mod consts;
 mod dump_config;
 mod environment;
@@ -31,7 +32,7 @@ mod error_server;
 mod event;
 mod file_watcher;
 mod introspect_local;
-mod parser;
+mod node;
 mod proxy;
 mod servers;
 mod udf_builder;
